% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_order.R
\name{create_order}
\alias{create_order}
\title{Buat Pemesanan dan Simpan ke Riwayat Transaksi}
\usage{
create_order(username, amount, item)
}
\arguments{
\item{username}{Nama pengguna}

\item{amount}{Total biaya pemesanan}

\item{item}{Nama item/barang/jasa}
}
\value{
Pesan status pemesanan
}
\description{
Fungsi untuk mensimulasikan pemesanan barang/jasa.
}
