% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapus_akun.R
\name{hapus_akun}
\alias{hapus_akun}
\title{Hapus Akun}
\usage{
hapus_akun(username, password)
}
\arguments{
\item{username}{Nama pengguna}

\item{password}{Kata sandi pengguna}
}
\value{
Pesan status
}
\description{
Fungsi untuk menghapus akun dan seluruh datanya (saldo, transaksi, laporan).
}
