% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lihat_saldo.R
\name{lihat_saldo}
\alias{lihat_saldo}
\title{Lihat Saldo Pengguna}
\usage{
lihat_saldo(username)
}
\arguments{
\item{username}{Nama pengguna}
}
\value{
Saldo pengguna atau pesan error
}
\description{
Menampilkan saldo terakhir dari pengguna.
}
