% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lihat_semua_user.R
\name{lihat_semua_user}
\alias{lihat_semua_user}
\title{Melihat Semua User yang Terdaftar (khusus untuk Owner)}
\usage{
lihat_semua_user(requester)
}
\arguments{
\item{requester}{Username pemanggil fungsi}
}
\value{
Daftar username atau pesan error
}
\description{
Fungsi ini akan menampilkan seluruh daftar user yang telah terdaftar di OVA-Pay.
Hanya bisa dijalankan oleh akun 'owner'.
}
