% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_user().R
\name{login_user}
\alias{login_user}
\title{Login User}
\usage{
login_user(username, password)
}
\arguments{
\item{username}{Nama pengguna}

\item{password}{Kata sandi}
}
\value{
Pesan status login
}
\description{
Fungsi untuk memverifikasi login user berdasarkan username dan password.
}
