% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pay().R
\name{pay}
\alias{pay}
\title{Melakukan Pembayaran Tagihan}
\usage{
pay(username, amount, keterangan = "")
}
\arguments{
\item{username}{Nama pengguna}

\item{amount}{Jumlah pembayaran}

\item{keterangan}{Deskripsi pembayaran}
}
\value{
Pesan status pembayaran
}
\description{
Fungsi untuk melakukan pembayaran dari saldo user.
}
