% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_user.R
\name{register_user}
\alias{register_user}
\title{Register User}
\usage{
register_user(username, password)
}
\arguments{
\item{username}{Nama pengguna.}

\item{password}{Kata sandi.}
}
\value{
Pesan status.
}
\description{
Fungsi untuk registrasi user baru.
}
