% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topup.R
\name{topup}
\alias{topup}
\title{Top Up Saldo User}
\usage{
topup(username, amount)
}
\arguments{
\item{username}{Nama pengguna}

\item{amount}{Jumlah yang akan ditambahkan}
}
\value{
Saldo terbaru
}
\description{
Fungsi untuk menambahkan saldo ke akun user
}
