% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer().R
\name{transfer}
\alias{transfer}
\title{Transfer Saldo ke User Lain}
\usage{
transfer(from_user, to_user, amount)
}
\arguments{
\item{from_user}{Nama pengirim}

\item{to_user}{Nama penerima}

\item{amount}{Jumlah saldo yang ditransfer}
}
\value{
Pesan status transfer
}
\description{
Fungsi untuk mentransfer saldo dari satu user ke user lain (sesama OVA-Pay).
}
