% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_bank.R
\name{transfer_bank}
\alias{transfer_bank}
\title{Transfer ke Rekening Bank}
\usage{
transfer_bank(username, amount, bank)
}
\arguments{
\item{username}{Nama pengguna}

\item{amount}{Jumlah uang yang ditransfer}

\item{bank}{Tujuan bank (karakter)}
}
\value{
Pesan status transfer
}
\description{
Melakukan simulasi transfer ke rekening bank dari saldo pengguna.
}
