% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validasi_user.R
\name{validasi_user}
\alias{validasi_user}
\title{Validasi User Terdaftar}
\usage{
validasi_user(username)
}
\arguments{
\item{username}{Nama pengguna}
}
\value{
TRUE jika terdaftar, FALSE jika tidak
}
\description{
Mengecek apakah user dengan username tertentu sudah terdaftar.
}
